Address  #11 with a larger number
This patch should minimize the number of false positives for EL injection.

It uses a minimum number for each factor and instead of one two random numbers each so that under really worst conditions one factor is around 100 and the other higher. That should suffice for 99,9999% of all cases (educated guess).